Split conv module into pack and unpack modules
